Add base class for Generators.
