+++ src/samovar/generators/random.py
@@ -5,34 +5,10 @@ import sys
 from samovar.ast import Assert, Retract
 from samovar.database import Database
 
-
-# Python 2/3
-try:
-    xrange = xrange
-except NameError:
-    xrange = range
-
-# Python 2/3
-try:
-    unicode = unicode
-except NameError:
-    unicode = str
+from .base import xrange, unicode, Event, BaseGenerator
 
 
-class Event(object):
-    def __init__(self, rule, unifier):
-        self.rule = rule
-        self.unifier = unifier
-
-    def to_json(self):
-        u = dict([(unicode(k), unicode(v)) for k, v in self.unifier.items()])
-        return [self.rule.to_json(), u]
-
-    def __str__(self):
-        return self.rule.format(self.unifier)
-
-
-class RandomGenerator(object):
+class RandomGenerator(BaseGenerator):
     def __init__(self, random, world, scenario, verbosity=0, sorted_search=True):
         self.random = random
         self.world = world
